+++ src/main/java/at/fhtw/lifequalvienna/service/ParkService.java
@@ -71,7 +71,7 @@ public class ParkService {
 
         double distance = geoService.distance(lon, lat, nearestStation.getLon(), nearestStation.getLat());
 
-        score.setExplanation(score.getExplanation() + ", which is " + distance + "m away. This results in a Park score of: ");
+        score.setExplanation(score.getExplanation() + ", which is " + (int) distance + "m away. This results in a Park score of: ");
 
 
         if(distance >= MAX_DISTANCE) {
@@ -81,7 +81,7 @@ public class ParkService {
             score.setScore(norm * 100);
         }
 
-        score.setExplanation(score.getExplanation() + score.getScore() + "%");
+        score.setExplanation(score.getExplanation() + (int) score.getScore() + "%");
 
         return score;
     }
