+++ src/main/java/at/fhtw/lifequalvienna/service/PoliceStationService.java
@@ -72,7 +72,7 @@ public class PoliceStationService {
 
         double distance = geoService.distance(lon, lat, nearestStation.getLon(), nearestStation.getLat());
 
-        score.setExplanation(score.getExplanation() + ", which is " + distance + "m away. This results in a Police-Station score of: ");
+        score.setExplanation(score.getExplanation() + ", which is " + (int) distance + "m away. This results in a Police-Station score of: ");
 
 
         if(distance >= MAX_DISTANCE) {
@@ -82,7 +82,7 @@ public class PoliceStationService {
             score.setScore(norm * 100);
         }
 
-        score.setExplanation(score.getExplanation() + score.getScore() + "%");
+        score.setExplanation(score.getExplanation() + (int) score.getScore() + "%");
 
         return score;
     }
